Break geometry module into explicit files
